feature/producto
agrego estos cambios parciales de el toast

+++ controladores/ProductoCtr.php
@@ -1,6 +1,7 @@
 <?php
 require_once 'models/ProductoMdl.php';
 require_once 'models/ProductoDAO.php';
+require_once 'controladores/ToastCtr.php';
 
 class ProductoCtr
 {
@@ -12,16 +13,44 @@ class ProductoCtr
         $module = isset($_GET['module']) ? $_GET['module'] : '';
         $action = isset($_GET['action']) ? $_GET['action'] : '';
         $id = isset($_GET['id']) ? $_GET['id'] : '';
+        $status = isset($_GET['status']) ? $_GET['status'] : "";
+        $toast = new ToastCtr();
+        if ($status == "error") {
+            $description = isset($_GET['description']) ? $_GET['description'] : "";
+            $toast->mostrarToast($status, $description);
+        }
         if ($module == 'productos') {
             switch ($action) {
                 case 'created':
-                    $this->create();
+                    if ($_SERVER["REQUEST_METHOD"] == "POST") {
+                        $status = isset($_GET['status']) ? $_GET['status'] : "";
+                        if ($status != "success") {
+                            $this->create();
+                        }
+                    } else {
+                        if ($status == "success") {
+                            $toast->mostrarToast("exito", "Producto creado");
+                        }
+                    }
                     break;
                 case 'deleted':
-                    $this->delete($id);
+                    if ($status != "success") {
+                        $this->delete($id);
+                    } else if ($status == "success") {
+                        $toast->mostrarToast("exito", "Producto eliminado");
+                    }
                     break;
                 case 'edited':
-                    $this->update($id);
+                    if ($_SERVER["REQUEST_METHOD"] == "POST") {
+                        $status = isset($_GET['status']) ? $_GET['status'] : "";
+                        if ($status != "success") {
+                            $this->update($id);
+                        }
+                    } else {
+                        if ($status == "success") {
+                            $toast->mostrarToast("exito", "Producto editado");
+                        }
+                    }
                     break;
             }
         }
@@ -42,12 +71,22 @@ class ProductoCtr
     }
     public function getProductoById($id)
     {
-        return $this->productoDAO->getProductoById($id);
+        $producto = $this->productoDAO->getProductoById($id);
+        if (is_string($producto)) {
+            $toast = new ToastCtr();
+            $toast->mostrarToast("error", "error al traer el producto", $producto);
+        }
+        return $producto;
     }
 
     public function getAllProductos()
     {
-        return $this->productoDAO->getAllProductos();
+        $productos = $this->productoDAO->getAllProductos();
+        if (is_string($productos)) {
+            $toast = new ToastCtr();
+            $toast->mostrarToast("error", "error al traer los productos", $productos);
+        }
+        return $productos;
     }
 
     public function getPantallaCreate()
